Move default catcher sprite to its own file
